Use snprintf at all times instead of sprintf

+++ src/support/strsec.h
@@ -1,8 +1,8 @@
 #include <stdlib.h>
 
-int strcpy_s(char *dest, int n, const char *src);
-int strcat_s(char *dest, int n, const char *src);
-int sprintf_s(char *dest, int n, const char *format, ...);
+int strcpy_s(char *dest, size_t n, const char *src);
+int strcat_s(char *dest, size_t n, const char *src);
+int sprintf_s(char *dest, size_t n, const char *format, ...);
 int _strlwr_s(char *dest, size_t sz);
 
 int _ultoa_s(unsigned long value, char *str, size_t sz, int radix);
